[0076] Split s7.c arithmetic operations into s7_scheme_number.c
