+++ devel/0076.md
@@ -1,0 +1,91 @@
+# [0076] 将 s7.c 中的算术运算代码拆分到 s7_scheme_number.c
+
+## 1. 相关文档
+- [dddd.md](dddd.md) - 任务文档模板
+- [0072.md](0072.md) - 从 s7.c 拆分基础函数和谓词函数
+- [0074.md](0074.md) - 完善 call/cc 支持：测试与底层代码拆分
+
+## 2. 任务相关的代码文件
+- `src/s7.c` - 被拆分的原始文件
+- `src/s7_scheme_number.c` / `src/s7_scheme_number.h` - 拆分后的目标文件
+- `xmake.lua` - 构建系统（本次不需要修改，采用条件包含模式）
+- `tests/scheme/s7-test.scm` - s7 回归测试
+
+## 3. 如何测试
+
+### 3.1 编译测试
+```bash
+xmake b goldfish
+```
+
+### 3.2 确定性测试（单元测试）
+```bash
+bin/gf tests/scheme/s7-test.scm
+```
+
+### 3.3 全量回归测试
+```bash
+bin/gf test --changed-since=main
+```
+
+## 4. 如何提交
+
+提交前执行以下最少步骤：
+
+```bash
+bin/gf test --changed-since=main
+```
+
+## 5. 2026-06-01 将 s7.c 中的算术运算通用版本拆分到 s7_scheme_number.c
+
+### 5.1 What
+
+将 `s7.c` 中从 `/* ---------------------------------------- add ---------------------------------------- */` 到 `g_add_i_random` 结束之间的算术运算通用版本代码（约 3500 行）拆分至独立的 `src/s7_scheme_number.c`，降低 `s7.c` 的复杂度。
+
+1. **新建文件**
+   - `src/s7_scheme_number.h`：条件包含头文件，包含必要的宏定义和函数声明
+   - `src/s7_scheme_number.c`：算术运算实现，包含 `add`/`subtract`/`multiply`/`divide`/`num_eq`/`less`/`greater`/`less_or_equal`/`greater_or_equal`/`random`/`integer_decode_float`/`integer_length` 等通用版本及辅助函数
+
+2. **迁移的主要函数类别**
+   - **四则运算通用版本**：`g_add`、`g_subtract`、`g_multiply`、`g_divide` 及其变体（`g_add_x1`、`g_subtract_x1`、`g_divide_by_2`、`g_invert_x` 等）
+   - **数值比较通用版本**：`g_num_eq`、`g_less`、`g_greater`、`g_less_or_equal`、`g_greater_or_equal` 及其变体
+   - **random 相关**：`g_random`、`g_random_1`、`g_random_f`、`g_random_i`、`g_add_i_random`
+   - **辅助函数**：`add_p_pp`、`subtract_p_pp`、`multiply_p_pp`、`divide_p_pp` 及其 wrapped/ppp 变体
+   - **bridge 函数**：`s7i_subtract_p_pp`、`s7i_multiply_p_pp`、`s7i_divide_p_pp` 等（从 s7.c 迁移到 s7_scheme_number.c）
+
+3. **采用条件包含模式**
+   - 与 `s7_continuation.c` 保持一致：在 `s7_scheme_number.h` 末尾通过 `#ifdef S7_SCHEME_NUMBER_IMPLEMENTATION` 条件包含 `s7_scheme_number.c`
+   - 在 `s7.c` 中 `#define S7_SCHEME_NUMBER_IMPLEMENTATION` 后 `#include "s7_scheme_number.h"`
+   - 不在 `xmake.lua` 中单独编译，保持与 s7.c 同一编译单元，避免内部宏替换问题
+
+4. **s7.c 中的变更**
+   - 移除约 3500 行算术运算实现代码
+   - 添加 `#define S7_SCHEME_NUMBER_IMPLEMENTATION` 和 `#include "s7_scheme_number.h"`
+   - 保留函数注册点（如 `sc->add_4 = make_function_with_class(...)` 等）
+
+### 5.2 Why
+
+`s7.c` 目前仍有约 8.8 万行代码，其中算术运算通用版本（`+` `-` `*` `/` `=` `<` `>` `<=` `>=` `random` 等）占约 3500 行，且功能高度内聚：
+- 编译时间长，任何小改动都触发全量重编译
+- 代码定位困难，四则运算、比较函数和 random 等混杂在 s7.c 的大段代码中
+- 已有 `s7_scheme_base.c`（基础算术快捷版本）和 `s7_scheme_inexact.c`（不精确数），继续拆分 numbers 模块可形成完整的算术运算子系统
+
+通过条件包含模式拆分，可将 `s7.c` 代码量减少约 3500 行，各模块职责清晰，且不引入大量 bridge 函数替换工作。
+
+### 5.3 How
+
+**整体策略**：采用与 `s7_continuation.c` 相同的条件包含模式，以函数为单位整体迁移。
+
+**具体步骤**：
+1. **识别代码边界**：`s7.c` 中从 `/* ---------------------------------------- add ---------------------------------------- */`（约 13977 行）到 `g_add_i_random` 结束（约 17479 行）
+2. **创建目标文件**：新建 `src/s7_scheme_number.h` 和 `src/s7_scheme_number.c`
+3. **迁移实现**：将函数体从 `s7.c` 剪切到 `s7_scheme_number.c`
+4. **创建头文件**：在 `s7_scheme_number.h` 中添加必要的宏定义和函数声明（主要供 `s7_internal_helpers.h` 中已有的 bridge 声明使用）
+5. **条件包含**：在 `s7_scheme_number.h` 末尾添加 `#ifdef S7_SCHEME_NUMBER_IMPLEMENTATION #include "s7_scheme_number.c" #endif`
+6. **更新 s7.c**：在剪切位置添加 `#define S7_SCHEME_NUMBER_IMPLEMENTATION` 和 `#include "s7_scheme_number.h"`
+7. **运行测试**：确保 `xmake b goldfish` 和 `bin/gf test --changed-since=main` 通过
+
+**注意事项**：
+- `add_p_pp` 等辅助函数仍被 s7.c 中 optimizer（约 42511 行之后）直接引用，条件包含模式下这些引用仍然有效
+- `s7_scheme_base.c`（独立编译单元）通过 `s7_internal_helpers.h` 中的 bridge 声明（如 `s7i_subtract_p_pp`）调用算术辅助函数，bridge 函数定义迁移到 `s7_scheme_number.c` 后仍保持链接可见性
+- 函数注册点（`make_function_with_class` 调用）保留在 s7.c 的初始化部分，不迁移
